Add GetReplicaPath to testhelper
Some of the Gitaly tests are run also with Praefect in front of them to ensure
they behavior works with Praefect as well. Some of these tests also peek into the
filesystem to perform assertations on the repository's state after running some
RPCs. This won't work anymore when Praefect begins rewriting relative paths as
the tests expect to find the repository in the directory indicated by the relative
path. To keep these tests working, this commit adds the GetReplicaPath helper that
returns the repository's replica path that Praefect has assigned the repository.
The helper checks whether the test is running with Praefect, and if so, connects to
Praefect to resolve the replica path and returns it. As the helper checks for
Praefect, it can be plugged in the existing tests and they'll work with and without
Praefect in front.