datastore: Revert use of materialized views
Revert the introduction of materialized views for `valid_primaries`. As
it turns out, the changes cause incompatibilities with Postgres 11,
which is still actively in use. Furthermore, the performance issues we
have seen have not been fully fixed with this change, and we do not yet
fully understand the root cause for this.

Changelog: fixed