operations: Fix missing feature flag tests for UserMergeBranch
In commit c0e4ef911 (operations: Always return structured errors from
Git2Go merges, 2021-09-21), we have removed the feature flag which
switched over the "merge" subcommand of gitaly-git2go to use gob for
serialization, and of course, this commit also removed the feature flag
from the tests. While at it, the change was a bit too eager and also
removed testing for the feature flag guarding structured errors in
UserMergeBranch, and so the old code isn't currently tested at all.

Reintroduce testing of the feature flag.