Generalize GetPartiallyAvailableRepositories for getting metadata
The only way to get a repository's metadata from Praefect is currently
the dataloss subcommand. However, the dataloss subcommand only works
if the repository is not fully replicated. This is not good as the admins
may be interested in the repository's metadata even if it is fully
replicated. This has become increasingly important with repository-specific
primaries and in the near future with Praefect generated relative paths.
To make it possible to get a repository's complete metadata always, this
commit generalizes the GetPartiallyAvailableRepositories and adds two
more methods for getting the metadata either by a repository's ID or by
it's (virtual_storage, relative_path). The new methods will be used in an
RPC that serves a new cli tool for getting the repository's metadata. The
query and the tests are shared between all of the methods to ensure they
stay in sync.