ci: Merge "proto" and "verify" jobs
We have two jobs which execute the "verify" and the "proto" targets,
respectively to check whether Go modules are tidy, whether the NOTICE
file is up-to-date, whether Protobuf files have been generated and to
run linting. The "verify" target does all the "proto" target already
does though, so the second job to explicitly call this target is
duplicating functionality we already have. The only additional thing it
brings to the table is that it will upload changed Protobuf files as
artifacts if this target fails.

Merge the "proto" job into the "verify" job to fix this duplication.