ci: Deduplicate definition of container image
While we have a global default image, this default image is not
parameterized with the Ruby and Go version. Because of this, we have to
redefine the image in multiple locations, while we don't in others. This
is confusing, and ultimately completely unnecessary: we can just add the
variables in the default image given that we also specify both Ruby and
Go versions as default variables anyway.

Remove the duplicate image definitions and instead parameterize the
default image.