ci: Merge postgres and test definitions
Nowadays, all tests use the Postgres database even when not using the
"test-with-praefect" target. As a result, each test job we have defined
includes both the "test_definition" and the "postgres_definition".

Remove this duplication by inlining the "postgres_definition" into the
"test_definition". While at it, also inline the "pgbouncer_definition"
into the only job which is using it.