ci: Let "coverage" job include test definitions
The "coverage" job is executing tests just as all the other targets we
have, but it doesn't include the "test_definition". Fix it to instead
include "test_definition" such that we can iterate more easily on all
test jobs at once. This also lets us get rid of the stage definition
given that it's already part of the test definition.