testhelper: Explicitly set all feature flags in feature sets
Feature sets only disable feature flags, but won't ever explicitly
enable a feature flag which is supposed to be active in the current
feature set. This works as expected in combination with feature flag
defaults, but is going to cause problems when we require feature flags
to be explicitly set in our tests.

Prepare for this future by always injecting all feature flags
explicitly.