Replace WaitForQueries with WaitForBlockedQuery
WaitForQueries is currently just checking the database has run the
expected number of queries as the last query from the given number of
connections. The tests that use this helper want to wait for a
query to be waiting on a lock. This ensures there is another
instance of the query actually blocked. This commit changes the function
to block until there is a given query blocked on a lock to better ensure
the correct case is hit. Additionally, the helper is not currently checking
the query is in the same database as the test is running. Parallel tests
running a matching query in a different database can cause this condition
to flake. This commit thus includes and additional check to ensure the query
is running in the same database as the current test.