Add DeriveReplicaPath function to praefectutil
Praefect will begin generating unique paths for the repositories to
avoid races with deleting and renaming repositories. Praefect assigns
a unique and permanent repository ID for each repository created on the
cluster. This unique ID can be used to derive unique paths for the
repositories. To avoid packing all repositories in a single directory,
the repository ID is hashed with SHA256 and the first four hex digits
are used as parent directories of the repository.

DeriveReplicaPath is placed in a new package to avoid import cycles between
Praefect and various gitaly services. Some of the Gitaly tests test what
happens if a file/directory already exists in the path where a repository
is about to be created. Such tests need to adjusts the relative paths when
Praefect is ran in front of them, which they'll need to use this function
for.