datastore: Introduction of the jackc/pgx
This change adds jackc/pgx driver as a replacement
for the lib/pq driver. To reduce code changes to
minimum we are using github.com/jackc/pgx/v4/stdlib
at first as it provides support of the standard
database/sql package, but still uses pgx underneath.

The configuration 'binary_parameters' has a different
name in pgx that is why we replace it, but it still does
the same thing.