commit: Remove unused arguments for `treeEntries()`
Before we've reimplemented recursive listings of tree entries to use
git-ls-tree(1), the `treeEntries()` helper function was used to list
entries both for recursive and non-recursive queries. But with the
feature flag removed, the helper is only ever used to do non-recursive
listings.

Remove the arguments which had been used to implement the recursion and
the recursive call itself.