Makefile: Show diff in no-changes
The no-changes Makefile target verifies that there are no changes in the
working tree, which is used by some CI jobs to verify whether generated
files are up-to-date or not. The target parses git-status(1) for this,
but this is not really all that helpful for us in case there _were_
changes: the caller wouldn't know what was changed, but only see the
failed Makefile target.

Improve the target by using `git diff --exit-code` instead: this is
simpler given we don't have to do post-processing via awk(1), and it
prints the diff such that callers know what has changed by just taking a
simple look at the CI job's output.