ssh: Fix overwriting shared gitaly-hooks binary
With the upcoming bundled Git changes, we'll start to see failures in
the ssh tests. The test failures are such that some tests seemingly
weren't able to find the Git executable anymore, but interestingly
enough output also contained references to a temporary directory which
ought to contain an "output" file. As it turns out, this output stems
from a testcase which writes its own gitaly-hooks script. So somehow,
this script is leaking into other tests. And given that this script also
contains the path to the Git binary hosted in our current temporary Git
execution env, it's not much of a surprise that subsequent tests cannot
find it anymore.

As it turns out, the root cause for this is the recent introduction of
using hardlinks for linking binaries which were built just-in-time into
place in a073666ab (testcfg: Try to fix ETXTBSY errors with
gitaly-git2go executable, 2021-11-03). Because the test is first
building gitaly-hooks and then overwrites the resulting binary again,
the end result is that we overwrite the shared gitaly-hooks binary.

Fix this issue by not building the gitaly-hooks binary anymore.
Furthermore, to help avoid such issues in the future,
`WriteExecutable()` as changed such that it will fail if the target
file exists already. This requires one more fix for merge tests, which
write the same script Git hook twice.