Fix go test report filename
In the Makefile, GIT_VERSION is overwritten when it is default. This
makes the junit filename no longer match what is specified in
gitlab-ci.yml. Instead here we are simplifying down to a static
filename. There shouldn't be any risk of the report being overwritten
since CI runs each permutation of the test matrix in separate
containers.