Makefile: Fix invalid usage of GIT_INSTALL_DIR
When installing Git via `make git`, then we have two variables which
guide where things will be installed to: GIT_PREFIX is user-controlled
and determines where Git should be installed to, and GIT_INSTALL_DIR is
the default value in case GIT_PREFIX is not set. So ultimately, Git will
always be installed into GIT_PREFIX.

The rules we have for installing Git use GIT_INSTALL_DIR as target
though, so when one installs Git with a manually-chosen GIT_PREFIX, then
the target mismatches the location where Git actually ends up. This used
to work alright given that we never did anything with the target except
for executing `make install` in the Git project, but with the recent to
introduce a bundled Git version we started to touch(1) the target. This
thus now fails.

Fix this by using GIT_PREFIX instead of GIT_INSTALL_DIR.

Changelog: fixed