Makefile: Do not create Git install directory manually
When installing Git via `make git`, we first clean up the
GIT_INSTALL_DIR by removing it completely. This is expected, as it is
the default location of the Git installation located in our own build
directory, and we want to have a tidy installation in there. After the
cleanup, we do recreate the target directory though, which is completely
unnecessary: the Makefile target would create it for us anyway, and in
case the default installation directory isn't used we wouldn't use that
directory anyway.

Stop creating the directory and let the Makefile target handle it for
us.