Makefile: Clarify GIT_INSTALL_DIR by renaming it
The GIT_INSTALL_DIR variable points into our own build directory and is
the default directory into which Git will be installed in case the user
didn't override GIT_PREFIX. The naming of this variable doesn't help at
all in underestanding what it does though and frequently leads to
confusion which of both variables should be used.

Rename the variable to GIT_DEFAULT_PREFIX to clarify what it does.