commandstatshandler: Do not modify global test logger variable
The commandstatshandler modifies the global testhelper variable which
keeps a reference to the function used to create a new test logger such
that we can stub out the log buffer when the logger is created in
`createNewServer()`. Depending on this global state is completely
unnecessary though: there's only a single user of `createNewServer()`,
so we can instead just pass in a logger.

Do so such that we can remove the global state altogether in a follow-up
commit.