Makefile: Only remove default Git prefix if installing into it
We have two different variables tracking the GIT_PREFIX: first the
GIT_PREFIX itself, which can be set by the user to point wherever the
user wants to install the compiled Git distribution to. And then we have
GIT_DEFAULT_PREFIX, which is the default value of GIT_PREFIX. There is
one special thing about GIT_DEFAULT_PREFIX though: upon installing Git,
we always remove the GIT_DEFAULT_PREFIX first regardless of whether
we're actually using the default prefix or not.

While it is clear that we _never_ want to remove GIT_PREFIX before
installing into it given that it may point e.g. to `/usr/local`, it's
questionable what the intent is in removing the default prefix first.
If I am about to guess, then we likely want to ensure that developers
never end up with Git installations of mixed versions given that the Git
version changes much more frequently on a developer machine, may it be
because of normal upgrades of the Git version or because one is in the
process of bisecting. As such, it is not too unlikely that the user will
end up with such mixed Git installations. And given that distros of
Gitaly will hopefully use proper packaging, we should assume that they
do not run such mixed installations either.

Even so, it does not make sense to delete the default prefix in case the
current prefix is different. So let's only delete it in case GIT_PREFIX
and GIT_DEFAULT_PREFIX are the same.