Makefile: Silence git-fsck's dangling object output
After cloning test repos, we replace their references with our own
vision of references by removing all loose ones and moving in a
pre-packed packed-refs file. As a result, we get a bunch of dangling
objects because we effectively move back the repository in time even
though we already downloaded all objects. While not bad per-se, it
creates quite some output when we run git-fsck(1) to check the repo for
consistency.

Silence these messages by adding the `--no-dangling` switch.