Makefile: Allow specifying test output format
With the introduction of gotestsum, the default output format has
changed to a new "short" summary. While this format is indeed nicer to
look at, it reports less information about what is going on and which
tests ran.

Add a new `TEST_FORMAT` variable which allows to change the formatting.
Note that this commit also drops the `-v` flag from `TEST_OPTIONS`: this
flag is not respected anymore because verbosity is now controlled by
gotestsum.