ci: Split up Ruby and dependency caches
We're currently using a single cache for both our build-time
dependencies like Git and libgit2, and for our Ruby Gems. Ideally, this
cache would also include our Go modules, but adding this in is busting
the size limitations.

Prepare for introducing a separate Go modules cache by splitting up the
Ruby and dependency caches. Like this, each job can include only the
caches it really needs, which reduces the time it takes to fetch and
decompress the archives. And furthermore, by splitting up the caches, we
reduce their respective sizes such that we can then easily introduce
another cache for Go modules, only.