praefect: Improve connectivity check test to not use timeout
The node connectivity check tests verify that the the check fails as
expected in case the Gitaly nodes are not reachable. We use a timeout to
do this, which isn't really required: we can just cancel the context
before passing it into the check, which will then cause us to fail
connecting to Gitaly nodes immediately.

Do so to get rid of the timeout.