supervisor: Refactor waiting for processes
Rewrite supervisor tests to not use timeouts anymore. Instead, we'll
always wait for the respective events to arrive. In the case where we
want to verify that the circuit breaker works as expected, we can assert
that the process takes at least `CrashWaitTime` until it gets respawned
instead of trying to connect to it with a timeout.

This allows us to greatly simplify much of the code and be race-free.