catfile: Convert adding processes to accept absolute deadline
When adding processes, we pass a time-to-live to the process cache such
that the process monitor will eventually kill processes which are older
than that time. Using a relative TTL makes the code harder than
necessary to test though given that it depends on how fast time
progresses.

Convert the function to instead accept an absolute expiry time. This
works completely the same as it did before, only that the expiry time is
computed in a different place.