catfile: Use absolute timings to test enforcement of time-to-live
The tests which exercise the time-to-live use relative durations right
now, which thus requires us to use sleeps between adding the processes
to the cache. This is unnecessary now that we can pass in absolute
expiry times.

Refactor the test to use absolute expiry times and get rid of the
sleeps.