catfile: Allow passing in a monitor ticker
The monitor ticker is currenly created by the catfile cache itself,
which means that it cannot be controlled from the outside. This makes it
hard to deterministically exercise the code in our tests.

Convert the interface to instead accept a `helper.Ticker` interface to
fix this. Note that compared to before, we must now manually reset the
ticker every time it fired.