catfile: Use manual ticker to test cache monitor
The cache monitor triggers in regular intervals to evict any processes
which are older than a given deadline. The test which exercises this
code is using a timer ticker, and as a result we need to sleep in the
test until the monitor may have been triggered.

Refactor the test to instead use a manual ticker and get rid of the
sleeps.