command: Use explicit context cancellation instead of sleep
The test which verifies that a command gets killed when the context is
cancelled is currently using a timeout for this: it spawns a sleep which
waits for three seconds, but the timeout is fired after one second.
While this hasn't caused any flakes to the best of my knowledge, it is
still unnecessarily flaky and wastes some time.

Refactor the test to instead sleep for one hour and use manual context
cancellation. Like this, there is no need for timeouts in the test given
that the test would be killed after 10 minutes anyway, and the test is
quicker to finish.