tests: Remove unneeded usage of `context.WithTimeout()`
There's a bunch of tests which use `context.WithTimeout()` without them
really requiring it. We do not want to benchmark how fast the system
under test is, and the timeuots may cause flaky tests on busy and/or
slow machines.

Remove usage of `context.WithTimeout()` from tests which don't need it.