git2go: Use gob for submodule sub-command invocation
So far we used JSON serialization to invoke submodule
sub-command of the git2go executable. We are moving
towards usage of the gob serialization to do the same
and use stdin and stdout of the process to get an input
and return the result of the execution. gob doesn't
require us to do mapping of the fields and allows
passing typed errors between the processes.

In order to support smooth migration we should support
both serialization formats. We can drop support of
the JSON in the next release.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3232

Changelog: changed