ci: Tie in PgBouncer as a service
The PgBouncer test job requires both Postgres and PgBouncer to run
concurrently. While the former runs as a service, the latter doesn't.
Instead, we run it manually and let it daemonize into the background via
the start script. This has the downside that we require PgBouncer to be
part of the build image, and furthermore it's not good practice.

Convert PgBouncer to instead run as a service.