remote: Refactor test for FindRemoteRootRef validation
Test test which exercises the validation logic for FindRemoteRootRef
requests is currently testing against two different validations at the
same time: if running without Praefect, then tests exercise only
Gitaly's validation, but if running with Praefect they also exercise how
Praefect validates the requests. The latter isn't any of the business of
this test though, and it waters down the test.

Disable usage of Praefect in this test such that we can properly assert
errors for equality again.