log: Fix upcoming cyclic dependency
The testhelper package needs to depend on the log package such that we
can properly configure Gitaly's logger. As a result, the log package
cannot import the testhelper package in any of its tests. We're about to
move the gRPC helper functions of the testassert package into the
testhelper package though, which are used by tests in the log package.

Prepare for this move and evade this cyclic dependency by not using
`testassert.ProtoEqual()` in the log test anymore. The case where we've
been using it works alright with `require.Equal()` anyway.