git: Disallow use of replace refs
By writing so-called replace refs into Git's reference database,
contents of objects may be replaced with contents of a different object.
While this can be useful in some contexts, it is a dangerous thing to
support on the server side: an adversary can easily hide an object's
original contents with a benign-looking version and thus sneak in a
malicious object into the repository.

Disallow use of replace refs globally to close down this potential
attack vector. While this is done for all Git commands, we only
demonstrate that this fix works via a catfile test.

Changelog: security