config: Omit writing values if they are empty
We are generation a configuration file when we
want to run praefect for the testing purposes.
The issue is that we are generating a file with
all values even if they are not set because we
are not using omitempty option for marshalling.
With the option set the setting won't be included
in the resulting file in case it has a default
value (0, "", nil, etc.).

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3304